+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Image container metadata
 - Acquire MRI Image Filter (combines MRI Signal, transform/resampling, noise)
 - Append metadata filter
+- Combine time-series images filter
+- Phase-magnitude filter
 
 ## [1.0.0] - 2020-10-16
 ### Added